refactor: dto ModelAttribute로 변경
