+++ src/main/java/com/woowacourse/momoimage/service/dto/ImageDto.java
@@ -1,12 +1,14 @@
-package com.woowacourse.momoimage.service;
+package com.woowacourse.momoimage.service.dto;
 
 import org.springframework.web.multipart.MultipartFile;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
+@Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class ImageDto {
